move ESB button to it's own file
